requirements.txt
-i https://pypi.org/simple
arctic==1.79.3
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
enum-compat==0.0.3
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
lxml==4.6.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
lz4==3.1.0; python_version >= '3.5'
mockextras==1.0.2
numpy==1.19.2; python_version >= '3.6'
pandas-datareader==0.9.0
pandas==1.1.3; python_full_version >= '3.6.1'
psycopg2==2.8.6
pymongo[zstd]==3.11.0
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tzlocal==2.1
urllib3==1.25.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
zstandard==0.14.0
